requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
exceptiongroup==1.2.0
h11==0.14.0
idna==3.4
numpy==1.26.2
opencv-python-headless==4.8.1.78
outcome==1.3.0.post0
packaging==23.2
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3.post1
pyzbar==0.1.9
requests==2.31.0
selenium==4.15.2
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
urllib3==2.1.0
webdriver-manager==4.0.1
wsproto==1.2.0
